+++ caveat/models/__init__.py
@@ -3,9 +3,9 @@ from .discrete.embed_conv import Conv
 from .discrete.lstm_discrete import LSTM_Discrete
 from .discrete.transformer_discrete import AttentionDiscrete
 from .embed import CustomDurationEmbedding
-from .sequence.conditional_lstm import ConditionalLSTM
-from .sequence.conditional_vae_lstm import CVAE_LSTM
-from .sequence.vae_lstm import VAE_LSTM
+from .sequence.cond_lstm import ConditionalLSTM
+from .sequence.cond_gen_lstm import CVAE_LSTM
+from .sequence.gen_lstm import VAE_LSTM
 
 library = {
     "VAE_Conv_Discrete": Conv,
