+++ tests/test_models.py
@@ -3,8 +3,8 @@ import torch
 from caveat.models.discrete.embed_conv import Conv
 from caveat.models.discrete.lstm_discrete import LSTM_Discrete
 from caveat.models.discrete.transformer_discrete import AttentionDiscrete
-from caveat.models.sequence.conditional_vae_lstm import CVAE_LSTM
-from caveat.models.sequence.vae_lstm import VAE_LSTM
+from caveat.models.sequence.cond_gen_lstm import CVAE_LSTM
+from caveat.models.sequence.gen_lstm import VAE_LSTM
 
 
 def test_conv_embed_cov_forward():
